Unify/Talk_To_Your_Data/requirements.txt
pandas==1.5.3
pandasai==2.0.32
streamlit==1.33.0
unifyai
pyyaml
